Rename repository to reflect split chunk fetching functionality
Updated the repository name to more accurately depict its primary function as a split chunk fetcher. This better aligns the repository's name with its current features and expected usage.
