+++ README.md
@@ -15,14 +15,14 @@ The Multi-Source Downloader is a highly efficient application crafted in Go. It 
 ## :wrench: Usage
 
 ```ruby
-multi-source-downloader [flags]
+split-fetcher [flags]
 
 Flags:
 
   -a, --assemble-only          // (Optional) Assemble part files only if true and --parts-dir and --manifest flags are passed
   -f, --decrypt-manifest       // (Optional) If true, decrypts the manifest file
   -d, --download-only          // (Optional) Download part files only if true
-  -h, --help                   // help for multi-source-downloader
+  -h, --help                   // help for split-fetcher
   -k, --keep-parts             // (Optional) Whether to keep the parts files after assembly
   -m, --manifest-file string   // (Required by --assemble-only) Manifest file (must be decrypted) to pass to the main function
   -c, --max-connections int    // (Optional) Controls how many parts of the file are downloaded at the same time.
@@ -44,26 +44,26 @@ Flags:
 To build the application, you need to execute the following command:
 
 ```ruby
-$ go build multi-source-downloader.go
+$ go build split-fetcher.go
 ```
 
-This command will compile the application and generate an executable binary named multi-source-downloader. This binary will be located in the same directory where the command was executed.
+This command will compile the application and generate an executable binary named split-fetcher. This binary will be located in the same directory where the command was executed.
 
 ## Running the Application
 
 To run the application directly without creating a binary, you can use the following command:
 
 ```ruby
-$ go run multi-source-downloader.go -s $sha_sums -u $url_file -k
+$ go run split-fetcher.go -s $sha_sums -u $url_file -k
 ```
 
 ## Using the Binary
 
-All examples provided in the following sections of this document use the multi-source-downloader binary to run the application. Once the binary is built, you can use it directly to execute the program without needing to use the go run command.
+All examples provided in the following sections of this document use the split-fetcher binary to run the application. Once the binary is built, you can use it directly to execute the program without needing to use the go run command.
 
 ## Examples:
 
-1. To run the multi-source-downloader application in its default mode. In this mode, the application is instructed to download a file from the URL that is provided as a parameter (in this case, $url_file).
+1. To run the split-fetcher application in its default mode. In this mode, the application is instructed to download a file from the URL that is provided as a parameter (in this case, $url_file).
 &nbsp;
       The application follows a few steps in default mode:
 
@@ -80,42 +80,42 @@ All examples provided in the following sections of this document use the multi-s
 ```ruby
 $ url_file="https://ftp.halifax.rwth-aachen.de/ubuntu-releases/23.04/ubuntu-23.04-netboot-amd64.tar.gz"
 
-$ ./multi-source-downloader -u $url_file
+$ ./split-fetcher -u $url_file
 ```
 
-2. To direct the multi-source-downloader to download a file from $url_file in 100 parts, using a maximum of 20 connections concurrently, and validates the integrity of the download using SHA sums from $shasums.
+2. To direct the split-fetcher to download a file from $url_file in 100 parts, using a maximum of 20 connections concurrently, and validates the integrity of the download using SHA sums from $shasums.
 
 ```ruby
 $ shasums="https://ftp.halifax.rwth-aachen.de/ubuntu-releases/23.04/SHA256SUMS"
 $ url_file="https://ftp.halifax.rwth-aachen.de/ubuntu-releases/23.04/ubuntu-23.04-netboot-amd64.tar.gz"
 
-$ ./multi-source-downloader -s $shasums -u $url_file -n 100 -c 20
+$ ./split-fetcher -s $shasums -u $url_file -n 100 -c 20
 ```
 
-3. To run the multi-source-downloader to download a file from $url_file, specified by -u, in 10 parts (-n 10), with a maximum of 5 connections at a time (-c 5). The parts are downloaded to the download_parts directory (-p download_parts), and the integrity of the downloaded file is verified using the SHA sums from $shasums (-s $shasums). The part files are kept after assembly (-k).
+3. To run the split-fetcher to download a file from $url_file, specified by -u, in 10 parts (-n 10), with a maximum of 5 connections at a time (-c 5). The parts are downloaded to the download_parts directory (-p download_parts), and the integrity of the downloaded file is verified using the SHA sums from $shasums (-s $shasums). The part files are kept after assembly (-k).
 
 
 ```ruby
-$ ./multi-source-downloader -s $shasums -u $url_file -n 10 -c 5 -d -p download_parts -k
+$ ./split-fetcher -s $shasums -u $url_file -n 10 -c 5 -d -p download_parts -k
 ```
 
-4. To use the multi-source-downloader to decrypt an encrypted manifest file, specified by $encrypted_manifest.
+4. To use the split-fetcher to decrypt an encrypted manifest file, specified by $encrypted_manifest.
 
 ```ruby
-encrypted_manifest="~/.config/.multi-source-downloader/ubuntu-23.04-netboot-amd64.manifest.51628721468495e921b639a4121e7342.json.enc"
+encrypted_manifest="~/.config/.split-fetcher/ubuntu-23.04-netboot-amd64.manifest.51628721468495e921b639a4121e7342.json.enc"
 
-$ ./multi-source-downloader -m $encrypted_manifest -f
+$ ./split-fetcher -m $encrypted_manifest -f
 ```
 
 &nbsp;
 **Note:** To decrypt the manifest file, access to the parts files is essential. These files, which are integral to the assembly of the final file, must be reachable for the application to operate properly. If the application cannot access these parts files, the decryption process will fail and the assembly of the final file will not be possible. Therefore, before you run the application, ensure that the parts files are accessible.
 
-5. The following example instructs the multi-source-downloader to assemble parts, stored in the parts_test directory and defined by the specified manifest file ($manifest), into a final file named output.tar.gz within the assembled directory.
+5. The following example instructs the split-fetcher to assemble parts, stored in the parts_test directory and defined by the specified manifest file ($manifest), into a final file named output.tar.gz within the assembled directory.
 
 ```go
-$ manifest="~/.config/.multi-source-downloader/ubuntu-23.04-netboot-amd64.manifest.51628721468495e921b639a4121e7342.json"
+$ manifest="~/.config/.split-fetcher/ubuntu-23.04-netboot-amd64.manifest.51628721468495e921b639a4121e7342.json"
 
-$ ./multi-source-downloader -m $manifest -p parts_test -o assembled/output.tar.gz -a
+$ ./split-fetcher -m $manifest -p parts_test -o assembled/output.tar.gz -a
 ```
 
 ## :zap: Fast & Concurrent
