+++ config/config_test.go
@@ -3,7 +3,7 @@ package config
 import (
 	"testing"
 
-	"github.com/campeon23/multi-source-downloader/logger"
+	"github.com/campeon23/split-fetcher/logger"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
 )
