+++ database/dbmanager/dbmanager.go
@@ -4,8 +4,8 @@ package dbmanager
 import (
 	"database/sql"
 
-	"github.com/campeon23/multi-source-downloader/database/initdb"
-	"github.com/campeon23/multi-source-downloader/logger"
+	"github.com/campeon23/split-fetcher/database/initdb"
+	"github.com/campeon23/split-fetcher/logger"
 )
 
 type DatabaseManager struct {
