+++ downloader/downloader.go
@@ -15,14 +15,14 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/campeon23/multi-source-downloader/config"
-	"github.com/campeon23/multi-source-downloader/database/initdb"
-	"github.com/campeon23/multi-source-downloader/encryption"
-	"github.com/campeon23/multi-source-downloader/fileutils"
-	"github.com/campeon23/multi-source-downloader/hasher"
-	"github.com/campeon23/multi-source-downloader/logger"
-	"github.com/campeon23/multi-source-downloader/manifest"
-	"github.com/campeon23/multi-source-downloader/utils"
+	"github.com/campeon23/split-fetcher/config"
+	"github.com/campeon23/split-fetcher/database/initdb"
+	"github.com/campeon23/split-fetcher/encryption"
+	"github.com/campeon23/split-fetcher/fileutils"
+	"github.com/campeon23/split-fetcher/hasher"
+	"github.com/campeon23/split-fetcher/logger"
+	"github.com/campeon23/split-fetcher/manifest"
+	"github.com/campeon23/split-fetcher/utils"
 
 	"github.com/google/uuid"
 	"github.com/gosuri/uiprogress"
