+++ encryption/encryption.go
@@ -14,10 +14,10 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/campeon23/multi-source-downloader/database/initdb"
-	"github.com/campeon23/multi-source-downloader/fileutils"
-	"github.com/campeon23/multi-source-downloader/logger"
-	"github.com/campeon23/multi-source-downloader/utils"
+	"github.com/campeon23/split-fetcher/database/initdb"
+	"github.com/campeon23/split-fetcher/fileutils"
+	"github.com/campeon23/split-fetcher/logger"
+	"github.com/campeon23/split-fetcher/utils"
 	"golang.org/x/crypto/argon2"
 )
