+++ fileutils/fielutils_test.go
@@ -4,7 +4,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/campeon23/multi-source-downloader/logger"
+	"github.com/campeon23/split-fetcher/logger"
 	"github.com/stretchr/testify/assert"
 )
