+++ fileutils/fileutils.go
@@ -12,9 +12,9 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/campeon23/multi-source-downloader/hasher"
-	"github.com/campeon23/multi-source-downloader/logger"
-	"github.com/campeon23/multi-source-downloader/utils"
+	"github.com/campeon23/split-fetcher/hasher"
+	"github.com/campeon23/split-fetcher/logger"
+	"github.com/campeon23/split-fetcher/utils"
 	"github.com/spf13/viper"
 )
