+++ manifest/manifest_test.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/campeon23/multi-source-downloader/logger"
+	"github.com/campeon23/split-fetcher/logger"
 	"github.com/stretchr/testify/assert"
 )
