+++ utils/utils.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/gosuri/uiprogress"
 
-	"github.com/campeon23/multi-source-downloader/logger"
+	"github.com/campeon23/split-fetcher/logger"
 )
 
 // Define a buffer pool globally to reuse buffers
